HDT: Update GATT tool to calculate throughput for LE COC data transfer.
The following changes are added in this patch -

1. Enhance both stack and tool to correctly send LE_SET_DEFAULT_HDT_PARAMETERS command.
2. Enhance tool to now calculate LE COC Data transfer throughput when data transfer initiated through tool.

CRs-Fixed: 4411045

+++ system/stack/include/hcimsgs.h
@@ -402,7 +402,7 @@ void btsnd_hci_ble_set_default_phy(uint8_t all_phys, uint8_t tx_phys, uint8_t rx
 
 void btsnd_hcic_le_set_hdt_default_parameters(uint8_t preferred_mic_length,
                                               uint8_t preferred_packet_format,
-                                              uint8_t preferred_acl_rates);
+                                              uint16_t preferred_acl_rates);
 
 void btsnd_hcic_ble_start_enc_v2(uint16_t handle, Octet8 rand, uint16_t ediv,
                               const Octet16& ltk, uint8_t hdt_mic_length, uint8_t enc_type);
